Move build scripts to Makefile in order to simplify development
Previously the only was to get files and directory
structure was to build it with RPM which took extra
time and efforts.

With this change the build happens in Makefile
which can be executed separately.

Example:

    make vendors

    builds vendors.d part of the package and places
    results in the build directory

    make core

    builds everything except core stuff

Usual workflow is now to make changes in code and
run make all to get structure that can be uploaded
on server or used as rpm source.